ml-items-challenge/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.23
requests==2.25.1
SQLAlchemy==1.4.4